# swagger.yml
openapi: 3.0.0
info:
  title: "Trail Management API"
  description: "API for user authentication, trail management, and feature management"
  version: "1.0.0"
servers:
  - url: "/api"
paths:
...
  /login:
    post:
      tags:
        - Authentication
      summary: "Log in a user"
      description: "Authenticate the user using their email and password."
      operationId: auth.login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: "password123"
      responses:
        "200":
          description: "Login successful"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  user:
                    type: object
                    properties:
                      user_id:
                        type: integer
                        example: 1
                      email:
                        type: string
                        example: user@example.com
        "401":
          description: "Invalid credentials"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /logout:
    post:
      tags:
        - Authentication
      summary: "Log out a user"
      description: "Log out the user and end their session using the email provided in the headers."
      operationId: auth.logout
      parameters:
        - name: email
          in: header
          required: true
          schema:
            type: string
            format: email
            example: grace@plymouth.ac.uk
            description: The email of the user to log out.
      responses:
        "200":
          description: "Logout successful"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User grace@plymouth.ac.uk logged out successfully."
        "400":
          description: "Missing email in headers"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing email in headers."
        "404":
          description: "No active session found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  #################### Trail Endpoints ####################
...
  /features:
    get:
      summary: Get All Features
      description: Fetch all features from the database and return them as a JSON response.
      operationId: features.read_all_features
      tags:
        - Features
      responses:
        '200':
          description: A list of features.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Feature'
        '500':
          description: An error occurred on the server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Add a New Feature
      description: Add a new feature to the database.
      operationId: features.add_feature
      tags:
        - Features
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                feature_name:
                  type: string
                  example: "Add new Feature"
      responses:
        '201':
          description: Feature successfully added.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Feature 'Waterfall' successfully added."
                  feature:
                    type: object
                    properties:
                      feature_name:
                        type: string
                        example: "Waterfall"
        '400':
          description: Validation error or feature already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /features/search:
    get:
      summary: Search Feature by Name
      description: Search for a feature by its name and return all trails associated with it.
      operationId: features.search_feature_by_name
      tags:
        - Features      
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trails associated with the feature.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrailWithFeatures"
        '400':
          description: Missing or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Feature not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /features/{current_feature_name}:
    put:
      summary: Update Feature by Name
      description: Update the name of an existing feature by searching for the current feature name.
      operationId: features.update_feature_by_name
      tags:
        - Features
      parameters:
        - name: current_feature_name
          in: path
          required: true
          schema:
            type: string
            example: "Waterfall"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                new_feature_name:
                  type: string
                  description: "The updated name for the feature."
                  example: "Scenic Waterfall"
      responses:
        "200":
          description: Feature name successfully updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Feature name successfully updated from 'Waterfall' to 'Scenic Waterfall'."
                  feature:
                    type: object
                    properties:
                      feature_name:
                        type: string
                        example: "Scenic Waterfall"
        "400":
          description: Validation error or duplicate name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Feature not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"              
  /features/{feature_name}:
    delete:
      summary: Delete Feature by Name
      description: Delete a feature from the database if it is not associated with any trail.
      operationId: features.delete_feature
      tags:
        - Features 
      parameters:
        - name: feature_name
          in: path
          required: true
          schema:
            type: string
            example: "Waterfall"
      responses:
        '200':
          description: Feature successfully deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Feature 'Waterfall' successfully deleted."
        '400':
          description: Feature is associated with trails and cannot be deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Feature not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


#################### Component Schemas ####################
components:
  schemas:
...
    TrailWithFeatures:
      type: object
      properties:
        trail_name:
          type: string
        trail_summary:
          type: string
        trail_description:
          type: string
        difficulty:
          type: string
        location:
          type: string
        length:
          type: number
          format: float
        elevation_gain:
          type: number
          format: float
        route_type:
          type: string
        waypoints:
          type: object
          properties:
            pt1:
              type: object
              properties:
                lat:
                  type: number
                  format: float
                long:
                  type: number
                  format: float
                desc:
                  type: string
            pt2:
              type: object
              properties:
                lat:
                  type: number
                  format: float
                long:
                  type: number
                  format: float
                desc:
                  type: string
            pt3:
              type: object
              properties:
                lat:
                  type: number
                  format: float
                long:
                  type: number
                  format: float
                desc:
                  type: string
        features:
          type: array
          items:
            type: object
            properties:
              feature_name:
                type: string
...
    Feature:
      type: object
      properties:
        feature_name:
          type: string
          example: "Waterfall"            
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "An error occurred."